Prepare for deprecation of dtypes

+++ scanpy/datasets/_datasets.py
@@ -50,7 +50,7 @@ def blobs(
         cluster_std=cluster_std,
         random_state=0,
     )
-    return AnnData(X, obs=dict(blobs=y.astype(str)))
+    return AnnData(X, obs=dict(blobs=y.astype(str)), dtype=X.dtype)
 
 
 @check_datasetdir_exists
@@ -179,7 +179,7 @@ def paul15() -> AnnData:
         clusters = f['cluster.id'][()].flatten().astype(int)
         infogenes_names = f['info.genes_strings'][()].astype(str)
     # each row has to correspond to a observation, therefore transpose
-    adata = AnnData(X.transpose())
+    adata = AnnData(X.transpose(), dtype=X.dtype)
     adata.var_names = gene_names
     adata.row_names = cell_names
     # names reflecting the cell type identifications from the paper
